Fixed the XML reader so it works on Silverlight. Made additional consistency changes between XNA and Silverlight versions
I think it would be best to focus work on Silverlight version. The only that that's a bit annoying is there isn't a real debugger for Silverlight, so it might be best to run code on XNA to debug, then transfer to Silverlight. However, there may be incompatibility issues... So it's best to try something in Silverlight first, debug in XNA, then finally move it to Silverlight.